+++ cmd/npv/app/reach.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"math/rand/v2"
 	"slices"
 	"strconv"
 
@@ -98,11 +97,6 @@ func runReach(ctx context.Context, w io.Writer) error {
 		return err
 	}
 
-	idEndpoints, err := getIdentityEndpoints(ctx, dynamicClient)
-	if err != nil {
-		return err
-	}
-
 	arr := make([]reachEntry, 0)
 
 	// process from-egress
@@ -158,9 +152,12 @@ func runReach(ctx context.Context, w io.Writer) error {
 				}
 			}
 			entry.Example = "-"
-			if v, ok := idEndpoints[p.Key.Identity]; ok {
-				i := rand.IntN(len(v))
-				entry.Example = v[i].GetName()
+			example, err := getIdentityExample(ctx, dynamicClient, p.Key.Identity)
+			if err != nil {
+				return err
+			}
+			if example != nil {
+				entry.Example = example.GetName()
 			} else {
 				idObj := identity.NumericIdentity(p.Key.Identity)
 				if idObj.IsReservedIdentity() {
@@ -242,9 +239,12 @@ func runReach(ctx context.Context, w io.Writer) error {
 				}
 			}
 			entry.Example = "-"
-			if v, ok := idEndpoints[p.Key.Identity]; ok {
-				i := rand.IntN(len(v))
-				entry.Example = v[i].GetName()
+			example, err := getIdentityExample(ctx, dynamicClient, p.Key.Identity)
+			if err != nil {
+				return err
+			}
+			if example != nil {
+				entry.Example = example.GetName()
 			} else {
 				idObj := identity.NumericIdentity(p.Key.Identity)
 				if idObj.IsReservedIdentity() {
@@ -277,12 +277,8 @@ func runReach(ctx context.Context, w io.Writer) error {
 	header := []string{"ROLE", "DIRECTION", "POLICY", "|", "IDENTITY", "NAMESPACE", "EXAMPLE-ENDPOINT", "|", "PROTOCOL", "PORT", "|", "BYTES:", "REQUESTS:", "AVERAGE:"}
 	return writeSimpleOrJson(w, arr, header, len(arr), func(index int) []any {
 		p := arr[index]
-		var protocol, port string
-		if p.WildcardProtocol {
-			protocol = "ANY"
-		} else {
-			protocol = u8proto.U8proto(p.Protocol).String()
-		}
+		protocol := u8proto.U8proto(p.Protocol).String()
+		var port string
 		if p.WildcardPort {
 			port = "ANY"
 		} else {
